infrastructure-in-azure/tests/requirements.txt
# Test dependencies for CI/CD pipeline
pytest>=7.0.0
testinfra>=6.0.0
requests>=2.28.0
PyYAML>=6.0
ansible>=6.0.0
ansible-lint>=6.0.0
